Replace C.ulonglong with C.uint64_t in Go fix #64
The problem comes from a discrepancy in types definitions between
Linux and Darwin: In the former, a `uint64_t` on the C side gets
mapped to a `C.ulong` on the go side, whereas in the latter it gets
mapped to a `C.ulonglong`. Ultimately, both types map to `unit64` on
Go side but one cannot use this directly.
Solution is simply to use `C.uint64_t` which provides the correct
mapping no matter what OS we use.