Include more precision in computation of phi.
Before we worked with `f64`, but the intention is to follow the mechanism performed in Cardano ledger. In there, they use the full 512 bits of the hash computation for the assignment of a participant, and then divide it by 2^512 to get a value in the range `[0, 1]`. We perform a slight performance improvement by computing the natural logarithm and a taylor expansion of `exp()` instead of the direct exponantiation.