Fix next aggregate verification key computation in Signer
The implementation needs to use a 'next_signers' field retrieved from the pending certificate in the Signer.
This is required to be able to compute correctly the next AVK in the message that is signed on the Signer.
Also the clerk from the multi signer of the Aggregator must use the correctly targeted verification keys:
currently it is always the verification keys for the current epoch.