Rename `SignerTickerImporter` to `SignersImporter`
This is more cardano agnostic and say better what it realy does (it does
not just import the ticker but also create a signer in the db if it does
not exist).
Also make one of its test more robust by using a BtreeMap in a assert_eq
instead of a HashMap (that does not have fixed order).