Split inform epoch of the certifier and the epoch service in state machine
They have two different purpose:
 - The inform epoch of the certifier is about cleaning up old data and
should be runned as soon as possible to tell signers that there's no
open message anymore.
 - The inform epoch of the epoch service is about preparing data that won't
change in the new epoch.
It should run as late as possible in order to ensure that everything is
available AND, since it will always fails because the needed data will
be missing, to ensure that the http server is ready when we are in a
'genesis' state.