Change open message creation in aggregator to use the state machine beacon
Before it was getting the beacon directly from the chain using the
ticker service.
This was problematic because its epoch could have evolved compared to the
one known by the stake machine, leading to computation using data from the
state machine known epoch but tagged as data of the next epoch.
IE: creating a open message for MithrilStakeDistribution(Epoch(13)) with
the avk of the epoch 12.