Modify load aggregator main scenario to try to send the 1st registration until the registration round is open
This is one the main source of flakiness of the load tester when running
with a lot of clients.
This happens because the epoch setting is the only thing that is needed
by signers to try to send a registration but the aggregator need a
little more time to be ready to open a registration round.
When running a lot of clients the database access in this interval are
slower meaning that the round take more time to open.