Remove all transactions usage in aggregator database layer
Else it may be sources of complicated bugs.

The change to the `ConnectionWithFullMutex` means that our methods can't
anymore lock the connection for their entire duration without the chance
of having a query executed in the middle of their duration.

We can still use transactions but in order to do so we must wrap all of
the transactions queries inside a single `connection.execute` so the
transaction applies only to its queries.

The existing transactions did not bring a lot of values (because we
don't have complexe queries) so it was easier to remove them all
together.