Correcly handle `Url::join` behaviour with trailing slash
The trailing slash is significant. Without it, the last path component is considered
to be a “file” name to be removed.

When `aggregator_url` is ending with a trailing slash (ex: https://whatever/aggregator/),
the `Url::join` with the endpoint works fine => https://whatever/aggregator/endpoint

In case there is no trailing slash in the `aggregator_url` (ex: https://whatever/aggregator),
the `Url::join` with the endpoint remove the last part of the url => https://whatever/endpoint