Refactor Certificate verification so all traits are defined by the client
This allow to reduce the number of re-export of `mithril-common` needed
since we can now use a `CerticateMessage` as our main certicate type
instead of using the `mithril_common::entities::Certificate` full type.
Meaning that we don't have to re-export crypto helpers or types anymore !

To do so we define a new `CertificateVerifier` trait that only have a
method that validate a chain and implement it with a new
`MithrilCertificateVerifier` struct that use internally the verifier
from `mithril-common`.
As with the `InternalCertificateRetriever` this add some complexity to
this specific part but the trade-off for it are worth it (as said
before: no crypto type re-export leading to far less mithril-common
dependency on the api level).