Explorer CertificateVerifier now take its client and cert as parameters
This give a cleaner hierarchy: the parent component now own those params
and the verifier only exploit them.
Before if the parent wanted to manipulate the certificate it would have
to get it from the verifier using a event callback.
This was unecessary cumbersome and also lead to the creation of more
wasm MithrilClient that necessary.