PR review: avoid using `RangeInclusive`
To be consistent with all usages or ranges for blocks, this remove the
last usage of a range inclusive to only use non-inclusive range.
The last case of RangeInclusive was a test that checked that an inverval
is contained in a block ranges sequence.

To do so it add a `contains(range)` to `BlockRangesSequence`.