Refactor transaction pruning to use two request instead of one
A first request obtains the highest `block_range_root.start` and a
second use it, and a given number of block to keep, as the prune
threshold.

We do that this instead of using a join to make it simple to test (else
the tests would need data on the two tables, our test framework doesn't
allow yet to do that without a lot of boilerplate).