Add Query system in mithril-persistence
Basically a `Query` a revamp provider without a reference to a
SQLite connection.
Instead of asking to implement a method that returns a connection it ask
to implement a method that return the `WhereCondition` to apply when
running the query.

The actual execution of the query is done by the `fetch` extension method
directly on the connection, allowing the following syntax:
```
let cursor =
connection.fetch(GetCardanoTransactionQuery::by_transaction_hash("hash"))?;
```

Also define two more extension methods to help when running sql queries:
* `fetch_one`: run `fetch` and call next one time, returning the result.
  Usefull when you only need one entity or when running CRUD queries
  (that are executed only after next is call on the cursor at least
  once).
* `fetch_and_collect`: run `fetch` and collect it's result in the
  desired containers as a classic collect does. Usefull when you want
  directly the result and don't care about the iterator.