Simplify CardanoTransaction signed entity type to `Epoch,BlockNumber`
Instead of `Epoch,ChainPoint`.

This is because we use this value as the upper bound of the blocks to
import and can't known easily the associated block hash and slot number
needed to constitute a ChainPoint (and we only need them for the lower
bound to allow Pallas to known the starting point).