Make SignableBuilder use a BlockNumber as target
This force to make the same change to the `TransactionImporter`,
`TransactionRetriever`,`BlockScanner` and the
`CardanoTransactionSnapshot`.

When a lower bound was required this is when we use a
`ChainPoint`(instead of a block number as before). To do this we added a
new condition to the `GetCardanoTransactionQuery` that get the
transactions with the highest block number in db (all transaction with a
same block number also share their slot number and block hash).

Co-authored-by: Sébastien Fauvel <sfauvel@users.noreply.github.com>
Co-authored-by: Damien Lachaume <dlachaume@users.noreply.github.com>