Simplify preloader setting by needing only a `security_parameter`
* Instead of asking for a `CardanoTransactionsSigningConfig` that
encompass the `security_parameter` plus add the signing `step`.

We can do this since the `step` is only needed when producing signatures
for `CardanoTransactions` signed entities, but the preloaded doesn't
produce any signatures and it's not a big deal if we store a little more
transactions and block range roots that what we would use if we wanted
to do a signature now.

* Make logs in the preloader one level higher.

Co-authored-by: Sébastien Fauvel <sfauvel@users.noreply.github.com>
Co-authored-by: Damien Lachaume <dlachaume@users.noreply.github.com>