Fix #1762: CardanoTransactionSnapshot block number is not certified
This is due to the request to retrieve the precomputed block range roots
in database using a exclusive bound comparaison instead of an inclusive,
leading to the exclusion of the last block range root from the retrieved
data.

This means that not only the prover was impacted but the protocol
message generation for the single & multi signatures too.

Co-authored-by: Sébastien Fauvel <sfauvel@users.noreply.github.com>