refactor: reduce hard fork test in CI to '8.9' version
Cardano node '8.12' has a bug that prevents it from creating
transactions when it is started with the Babbage era.

The error is due to the addition of parameters to the PlutusV2 cost
model which creates an error when creating a transaction with
the cardano-cli. This is a known issue that should be fixed with the next
release.