Make aggregator prover fail when trying to add key to the merkle tree
Before it allowed to add new keys to the merkle tree, making it create
proofs based on a different merkle root than the one signed in the last
certificate.

Even more worrisome this merkle tree would be then returned modified
to its mkpool, making this corruption persist until the pool is refreshed.