Change block range computation to use a provided upper bound
Instead of computing it from the database.

The problem with using bounds entirely computed from database is that it
can't accound for blocks with empty transactions, something rare but not
impossible in the mainnet.
Ie:
* There's some Transactions for each blocks in interval (15..=29): a
  block range root is computed for the interval (15..=29)
* There's some Transactions for each blocks in interval (15..=28), no
  transaction for block 29:
  today no block range root are computed but we should compute one since
  the absence of tx for that block is the actual chain state.

An alternative would be to store the blocks in the db so we could
compute the upper bound based on blocks not on transactions.