Use a RangeInclusive for `BlockRange::all_block_ranges_in`
Since all callers will now ask for an inclusive range instead of an
exclusive one like before.

This allow the TransactionImporter to be called with a block number
that's the strict end of a block range (ie: 44 for block range 30..45)
without the need to do a `+ 1` when calling `BlockRange::all_block_ranges_in`.