Increase e2e aggregator security parameter for CardanoTransactions
From `1` to `15`.

With only `1` the chain reader used to parse transactions from the chain
would switch to 'agency' mode since we are really close to the tip of
the chain.
In this mode the client/server relationship is reversed and our code
must wait a message from the cardano node in order to continue polling
blocks.
But the chain reader don't handle that case when we are setting the
starting chain point, leading to a recuring error that would block all
subsenquent calls.

So increasing the security parameter allow us to avoid this problem by
never be so close to the tip than we switch mode.