fix: last block read by Chain Reader in Block Streamer is strictly above the limit
In the Chain Reader Block Streamer, the Chain Reader reads the block strictly
above the threshold instead of stopping as soon as the threshold is reached.
When the Chain Reader is at the tip of the chain, the underlying chain sync client
switches agency and it can not find a new intersection then.
The result is that the block which was read too early was never recorded in the store,
which caused a wrong computation of some of the block range roots and thus of the protocol message.