Make aggregator multi signer able to verify signature issued for the next epoch
Meaning that the `EpochService` must store it (trivial since we already
built one to compute the next avk).

Also in order to be able to verify sigend_message of single signatures
the multisigner verify methods had to change to take string instead of
protocol message (we can't use the `AsMessage` trait as this will
prevent the `MultiSigner` from being made into an object).