Use sqlite store in buffered certifier tests
In order to remove the InMemory store.

Doing so required to be more precise with the tests signatures used. IE
most"register_signature" tests needs a single signature that have a
signed message but when reading them from the store the store doesn't
keep the signed message leading to problems with some assert_eq.