fix: stm tests run time regression
Merge of #2282 changed the feature set used by `mithril-client-wasm`,
including the activation of the `num-integer-backed` feature from the
`mithril-client` crate.
With feature unification this have the side effect of enabling this
slower backend for everyone when you run tests at the workspace level.

Removing this feature in client wasm is no-op since `mithril-stm`
enforce it when building to wasm or windows.