fix(common): don't compile type-tag in wasm
As it's incompatible.

This means that in wasm we can't serialize/deserialize a field that
contains a generics bound to type tagged trait (only `Artifact` trait
is impacted).
We only have one use case for this rigth now, storing artifacts in a
generic database records field, which is not compiled to wasm so this is
acceptable.