fix: ledgers and volatiles naming scheme were reversed in tests
This have no real impact, as those were only handled as pure strings,
but this allow to be coherent with the real naming scheme in the cardano
database.

Ledger are named after a number (a slot number)
Volatile files are named with the following format: `blocks-xxxx.data`,
were `xxxx` is a number.

Co-authored-by: Damien Lachaume <dlachaume@users.noreply.github.com>