refactor(node-db): move dummies to a `fake_data` module
This allow to keep a clean separation between production and test code
even when the later is available publicly.
Before this separation was made by using a feature flag, but we now want
to avoid features flags as mush as possible since it reduce ability for
the compiler to make cached build artifacts.