feat(common): impl `Default` to `APIVersionProvider`
This is useful for context when the `APIVersionProvider` don't need to
return response that may evolve at runtime, such as with the
aggregator-client in the mithril-client.

This also enable the new shared aggregator-client to have a default api
version provider, that will always use the base `openapi.yml`, but
provide a way to use another provider in its builder (for the aggregator
and the client that will fetch the version discriminant based on the
chain).