feat(client): add re-export of `mithril-common` `Dummy` trait
To allow usage without needing an import of common.

Note: reexport of `fake_data` and `fake_keys` were considered as well
but were discarded in order to maintain ability to changes them easily
without having to worry about breaking `mithril-client` library.
This problem is less relevant for the `Dummy` trait as it's really
simple and don't risk to change anytime soon.