fix(e2e): only update protocol parameters on leader aggregator
Since now the follower read the network config from the leader, this
means that the update of the protocol parameters is now a responsability
of the leader only.

This lead to flakiness because this step was restarting all aggregators,
and sometimes the follower started before the leader and had a error
when it executed its handle discrepencies because the leader http server
was still down.