feat(common): make cardano transactions signing config optional in `SignedEntityConfig`
This make the network resilient to a removal of this configuration at
the expense of moving the error handling to nodes 'beacon to sign'
computations (done in their state machines).

Notes: leader aggregator will fails directly at startup if an incoherent
configuration, which enable cardano transactions but don't specifies the
associated configuration, is provided.