test(e2e): avoid using a script in version parsing tests to make nix happy
This replace the `test_version_output_parsing` test with a simpler test
that just check a `std::process::Command` output instead of creating a
script to simulate a full version retrieval.

The check step in nix, that run `cargo test`, always failed against
`test_version_output_parsing`, it was able to create the script but not
to run it, failing with `No such file or directory (os error 2)`.

Investigations could not find the issue, it may be related to how nix
mange symlinks, but we don't know what's really going on.